+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Fixed
 
-- Keep extension commands available while the new-version changelog notification
-  is awaiting user interaction ([#60](https://github.com/micheletriaca/fast-sfdc/issues/60)).
+- Keep extension commands available while the new-version changelog notification is awaiting user interaction ([#60](https://github.com/micheletriaca/fast-sfdc/issues/60)).
+- Show Apex class and trigger test coverage with source-version checks.
+
 
 ## 2.0.2